修改提交人物和人物草稿的接口

+++ handlers/people/people_description.py
@@ -13,6 +13,7 @@ class CommitPeopleDescriptionHandler(tornado.web.RequestHandler):
         uploader = self.get_argument("username")
         time_stamp = self.get_argument("time_stamp")
         draft_people_id = self.get_argument("draft_people_id")
+        people_id = self.get_argument("people_id")
         description_text = self.get_argument("description_text")
         data = {}
         try:
@@ -20,11 +21,18 @@ class CommitPeopleDescriptionHandler(tornado.web.RequestHandler):
             people_description_id = mdb.insert_people_description(uploader=uploader,
                                                                   time_stamp=time_stamp,
                                                                   description_text=description_text,
-                                                                  draft_people_id=draft_people_id)
-            # 更新数据表
-            mdb.update_draft_people_description_id(draft_people_id=draft_people_id,
-                                                   time_stamp=time_stamp,
-                                                   description_id=people_description_id)
+                                                                  draft_people_id=draft_people_id,
+                                                                  people_id=people_id)
+            if draft_people_id is not None:
+                # 更新数据表
+                mdb.update_draft_people_description_id(draft_people_id=draft_people_id,
+                                                       time_stamp=time_stamp,
+                                                       description_id=people_description_id)
+            elif people_id is not None:
+                # 更新数据表
+                mdb.update_people_description_id(people_id=people_id,
+                                                 time_stamp=time_stamp,
+                                                 description_id=people_description_id)
         except BaseException as e:
             data['code'] = -1
             data['msg'] = "commit people description error"
@@ -32,7 +40,8 @@ class CommitPeopleDescriptionHandler(tornado.web.RequestHandler):
         else:
             # 删除草稿中的数据
             try:
-                mdb.delete_draft_people_description(draft_people_id)
+                if draft_people_id is not None:
+                    mdb.delete_draft_people_description(draft_people_id)
             except BaseException as e:
                 logging.exception(e)
             else:
@@ -50,7 +59,7 @@ class UpdatePeopleDescriptionHandler(tornado.web.RequestHandler):
         description_text = self.get_argument("description_text")
         data = {}
         try:
-            # 更新数据到people_event表
+            # 更新数据到people_description表
             mdb.update_people_description(uploader=uploader,
                                           time_stamp=time_stamp,
                                           description_text=description_text,
